backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
yfinance
openai>=1.0.0
python-dotenv
python-jose